langgraph-agents/requirements.txt
# LangGraph Multi-Agent System Requirements
# Updated to latest versions (December 2025)

# Core LangGraph dependencies (all 1.0+)
langgraph>=1.0.5
langchain>=1.1.2
langchain-core>=1.1.1
langchain-anthropic>=1.2.0

# LLM providers
anthropic>=0.75.0
langchain-openai>=0.3.0
openai>=1.50.0
langchain-google-genai>=2.1.0

# HTTP API server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.0.0

# Database and caching
redis>=6.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0  # Async SQLAlchemy for config service

# Async HTTP client
httpx>=0.28.0

# Observability
langsmith>=0.4.56
sentry-sdk[fastapi,httpx]>=2.19.0

# Modal SDK for sandbox execution
modal>=1.2.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.12.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Linting and type checking
ruff>=0.8.0
mypy>=1.13.0
types-redis>=4.6.0

# PostgreSQL checkpointing for state persistence
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.1.0

# SDK for client integration (not the licensed server)
langgraph-sdk>=0.3.0
